Clarified connection between the tutorial and RxJS
The README.md includes a clear statement that this is a tutorial about Rx, and explains why it's titled "Functional Programming in JavaScript." That bit was left out or cut out from the tutorial proper, which means that aside from the directory name and one brief mention just before exercise 28 (which is specific to Observables and not to the functions implemented), Rx is not mentioned at all.

Additionally, the tutorial is ambiguous about the nature of the functions. The brief related video tutorials at Egghead.io make it very clear that `.map()`, `.filter()`, and `.reduce()` are native to JS and that the rest of the functions explored are found within the RxJS library. That is never made explicit here, and many students run the risk of misunderstanding the purpose of "implementing" the various methods (e.g., are they modifying existing methods, creating new ones for their own use, or looking at the guts to better understand how they work?). It may be worthwhile addressing this throughout when each function is dealt with, but the initial clarification will help mitigate that confusion.