Bugfix in Answer 14, missing filter on height
Same error of previous commit for Exercise 12:

Proposed answer does not implement correctly the exercise "Use concatMap() to retrieve id, title, and 150x200 box art url for every video" because it does not take into account boxart.height. It work with proposed data just because there is no video with a 150xH boxart with H != 200.